Added support for VCR cassettes.
