+++ .github/workflows/unit-tests.yaml
@@ -35,7 +35,7 @@ jobs:
 
       - name: Run unit tests
         run: |
-          pytest tests
+          pytest tests --vcr-record=none
         env:
           NOTION_AUTH_TOKEN: ${{ secrets.NOTION_AUTH_TOKEN }}
           NOTION_TEST_AREA: ${{ secrets.NOTION_TEST_AREA }}
