+++ .yamllint
@@ -1,8 +1,13 @@
 ---
+extends: default
+
 yaml-files:
   - '*.yaml'
   - '*.yml'
   - '.yamllint'
+
+ignore: |
+  /tests/cassettes/
 
 rules:
   braces: enable
