+++ requirements/test.txt
@@ -1,4 +1,5 @@
 -r core.txt
 
 pytest==7.1.1
+pytest-vcr==1.0.2
 coverage==6.3.2
