+++ tests/cassettes/test_create_block.yaml
@@ -1,0 +1,73 @@
+interactions:
+- request:
+    body: '{"children": [{"type": "divider", "object": "block", "has_children": false,
+      "archived": false, "divider": {}}]}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '111'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: PATCH
+    uri: https://api.notion.com/v1/blocks/250662d9783d483cb131341ed479f96c/children
+  response:
+    content: '{"object":"list","results":[{"object":"block","id":"d87b1d70-87c4-457d-ae6d-5b9f73837fe4","created_time":"2022-04-15T15:38:00.000Z","last_edited_time":"2022-04-15T15:38:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":false,"type":"divider","divider":{}}],"next_cursor":null,"has_more":false}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: GET
+    uri: https://api.notion.com/v1/blocks/d87b1d70-87c4-457d-ae6d-5b9f73837fe4
+  response:
+    content: '{"object":"block","id":"d87b1d70-87c4-457d-ae6d-5b9f73837fe4","created_time":"2022-04-15T15:38:00.000Z","last_edited_time":"2022-04-15T15:38:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":false,"type":"divider","divider":{}}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: DELETE
+    uri: https://api.notion.com/v1/blocks/d87b1d7087c4457dae6d5b9f73837fe4
+  response:
+    content: '{"object":"block","id":"d87b1d70-87c4-457d-ae6d-5b9f73837fe4","created_time":"2022-04-15T15:38:00.000Z","last_edited_time":"2022-04-15T15:38:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":true,"type":"divider","divider":{}}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+version: 1
