+++ tests/cassettes/test_delete_block.yaml
@@ -1,0 +1,78 @@
+interactions:
+- request:
+    body: '{"children": [{"type": "code", "object": "block", "has_children": false,
+      "archived": false, "code": {"text": [{"type": "text", "plain_text": "test_delete_block",
+      "text": {"content": "test_delete_block"}}], "caption": [], "language": "plain
+      text"}}]}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '249'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: PATCH
+    uri: https://api.notion.com/v1/blocks/250662d9783d483cb131341ed479f96c/children
+  response:
+    content: '{"object":"list","results":[{"object":"block","id":"18decb3f-8465-46d0-ade4-7608b258f3f4","created_time":"2022-04-15T15:38:00.000Z","last_edited_time":"2022-04-15T15:38:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":false,"type":"code","code":{"caption":[],"language":"plain
+      text","text":[{"type":"text","text":{"content":"test_delete_block","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"test_delete_block","href":null}]}}],"next_cursor":null,"has_more":false}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: DELETE
+    uri: https://api.notion.com/v1/blocks/18decb3f846546d0ade47608b258f3f4
+  response:
+    content: '{"object":"block","id":"18decb3f-8465-46d0-ade4-7608b258f3f4","created_time":"2022-04-15T15:38:00.000Z","last_edited_time":"2022-04-15T15:38:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":true,"type":"code","code":{"caption":[],"language":"plain
+      text","text":[{"type":"text","text":{"content":"test_delete_block","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"test_delete_block","href":null}]}}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: GET
+    uri: https://api.notion.com/v1/blocks/18decb3f-8465-46d0-ade4-7608b258f3f4
+  response:
+    content: '{"object":"block","id":"18decb3f-8465-46d0-ade4-7608b258f3f4","created_time":"2022-04-15T15:38:00.000Z","last_edited_time":"2022-04-15T15:38:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":true,"type":"code","code":{"caption":[],"language":"plain
+      text","text":[{"type":"text","text":{"content":"test_delete_block","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"test_delete_block","href":null}]}}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+version: 1
