+++ tests/cassettes/test_missing_property.yaml
@@ -1,0 +1,117 @@
+interactions:
+- request:
+    body: '{"parent": {"type": "page_id", "page_id": "250662d9-783d-483c-b131-341ed479f96c"},
+      "title": [{"type": "text", "plain_text": "tests/test_orm.py::test_missing_property
+      (setup)", "text": {"content": "tests/test_orm.py::test_missing_property (setup)"}}],
+      "properties": {"Name": {"type": "title", "title": {}}, "Index": {"type": "number",
+      "number": {"format": "number"}}, "Notes": {"type": "rich_text", "rich_text":
+      {}}, "Complete": {"type": "checkbox", "checkbox": {}}, "Due Date": {"type":
+      "date", "date": {}}, "Tags": {"type": "multi_select", "multi_select": {"options":
+      []}}}}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '575'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: POST
+    uri: https://api.notion.com/v1/databases
+  response:
+    content: '{"object":"database","id":"f04a3545-e1cd-4592-ae5a-86a1b96dd64d","cover":null,"icon":null,"created_time":"2022-04-15T15:39:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_time":"2022-04-15T15:39:00.000Z","title":[{"type":"text","text":{"content":"tests/test_orm.py::test_missing_property
+      (setup)","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"tests/test_orm.py::test_missing_property
+      (setup)","href":null}],"properties":{"Tags":{"id":"TPRA","name":"Tags","type":"multi_select","multi_select":{"options":[]}},"Due
+      Date":{"id":"Tg%3CY","name":"Due Date","type":"date","date":{}},"Notes":{"id":"gDRz","name":"Notes","type":"rich_text","rich_text":{}},"Index":{"id":"qnUx","name":"Index","type":"number","number":{"format":"number"}},"Complete":{"id":"%7Cym%5C","name":"Complete","type":"checkbox","checkbox":{}},"Name":{"id":"title","name":"Name","type":"title","title":{}}},"parent":{"type":"page_id","page_id":"250662d9-783d-483c-b131-341ed479f96c"},"url":"https://www.notion.so/f04a3545e1cd4592ae5a86a1b96dd64d","archived":false}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: '{"parent": {"type": "database_id", "database_id": "f04a3545-e1cd-4592-ae5a-86a1b96dd64d"},
+      "properties": {}}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '108'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: POST
+    uri: https://api.notion.com/v1/pages
+  response:
+    content: '{"object":"page","id":"dd409a96-e5d6-489d-8403-009cf52ebe4a","created_time":"2022-04-15T15:39:00.000Z","last_edited_time":"2022-04-15T15:39:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"cover":null,"icon":null,"parent":{"type":"database_id","database_id":"f04a3545-e1cd-4592-ae5a-86a1b96dd64d"},"archived":false,"properties":{"Tags":{"id":"TPRA","type":"multi_select","multi_select":[]},"Due
+      Date":{"id":"Tg%3CY","type":"date","date":null},"Notes":{"id":"gDRz","type":"rich_text","rich_text":[]},"Index":{"id":"qnUx","type":"number","number":null},"Complete":{"id":"%7Cym%5C","type":"checkbox","checkbox":false},"Name":{"id":"title","type":"title","title":[]}},"url":"https://www.notion.so/dd409a96e5d6489d8403009cf52ebe4a"}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: '{"properties": {"Name": {"type": "title", "title": [{"type": "text", "plain_text":
+      "MissingProp", "text": {"content": "MissingProp"}}]}}}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '137'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: PATCH
+    uri: https://api.notion.com/v1/pages/dd409a96e5d6489d8403009cf52ebe4a
+  response:
+    content: '{"object":"page","id":"dd409a96-e5d6-489d-8403-009cf52ebe4a","created_time":"2022-04-15T15:39:00.000Z","last_edited_time":"2022-04-15T15:39:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"cover":null,"icon":null,"parent":{"type":"database_id","database_id":"f04a3545-e1cd-4592-ae5a-86a1b96dd64d"},"archived":false,"properties":{"Tags":{"id":"TPRA","type":"multi_select","multi_select":[]},"Due
+      Date":{"id":"Tg%3CY","type":"date","date":null},"Notes":{"id":"gDRz","type":"rich_text","rich_text":[]},"Index":{"id":"qnUx","type":"number","number":null},"Complete":{"id":"%7Cym%5C","type":"checkbox","checkbox":false},"Name":{"id":"title","type":"title","title":[{"type":"text","text":{"content":"MissingProp","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"MissingProp","href":null}]}},"url":"https://www.notion.so/MissingProp-dd409a96e5d6489d8403009cf52ebe4a"}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: DELETE
+    uri: https://api.notion.com/v1/blocks/f04a3545e1cd4592ae5a86a1b96dd64d
+  response:
+    content: '{"object":"block","id":"f04a3545-e1cd-4592-ae5a-86a1b96dd64d","created_time":"2022-04-15T15:39:00.000Z","last_edited_time":"2022-04-15T15:39:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":true,"type":"child_database","child_database":{"title":"tests/test_orm.py::test_missing_property
+      (setup)"}}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+version: 1
