+++ tests/cassettes/test_page_icon.yaml
@@ -1,0 +1,113 @@
+interactions:
+- request:
+    body: '{"parent": {"type": "page_id", "page_id": "250662d9-783d-483c-b131-341ed479f96c"},
+      "properties": {"title": {"type": "title", "title": [{"type": "text", "plain_text":
+      "tests/test_records.py::test_page_icon (setup)", "text": {"content": "tests/test_records.py::test_page_icon
+      (setup)"}}]}}}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '288'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: POST
+    uri: https://api.notion.com/v1/pages
+  response:
+    content: '{"object":"page","id":"746c707a-24ef-44f8-867d-17bf7d4371b7","created_time":"2022-04-15T15:39:00.000Z","last_edited_time":"2022-04-15T15:39:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"cover":null,"icon":null,"parent":{"type":"page_id","page_id":"250662d9-783d-483c-b131-341ed479f96c"},"archived":false,"properties":{"title":{"id":"title","type":"title","title":[{"type":"text","text":{"content":"tests/test_records.py::test_page_icon
+      (setup)","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"tests/test_records.py::test_page_icon
+      (setup)","href":null}]}},"url":"https://www.notion.so/tests-test_records-py-test_page_icon-setup-746c707a24ef44f8867d17bf7d4371b7"}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: '{"icon": {"type": "emoji", "emoji": "\u2603\ufe0f"}}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '52'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: PATCH
+    uri: https://api.notion.com/v1/pages/746c707a24ef44f8867d17bf7d4371b7
+  response:
+    content: "{\"object\":\"page\",\"id\":\"746c707a-24ef-44f8-867d-17bf7d4371b7\",\"created_time\":\"2022-04-15T15:39:00.000Z\",\"last_edited_time\":\"2022-04-15T15:39:00.000Z\",\"created_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"last_edited_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"cover\":null,\"icon\":{\"type\":\"emoji\",\"emoji\":\"\u2603\uFE0F\"},\"parent\":{\"type\":\"page_id\",\"page_id\":\"250662d9-783d-483c-b131-341ed479f96c\"},\"archived\":false,\"properties\":{\"title\":{\"id\":\"title\",\"type\":\"title\",\"title\":[{\"type\":\"text\",\"text\":{\"content\":\"tests/test_records.py::test_page_icon
+      (setup)\",\"link\":null},\"annotations\":{\"bold\":false,\"italic\":false,\"strikethrough\":false,\"underline\":false,\"code\":false,\"color\":\"default\"},\"plain_text\":\"tests/test_records.py::test_page_icon
+      (setup)\",\"href\":null}]}},\"url\":\"https://www.notion.so/tests-test_records-py-test_page_icon-setup-746c707a24ef44f8867d17bf7d4371b7\"}"
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: GET
+    uri: https://api.notion.com/v1/pages/746c707a-24ef-44f8-867d-17bf7d4371b7
+  response:
+    content: "{\"object\":\"page\",\"id\":\"746c707a-24ef-44f8-867d-17bf7d4371b7\",\"created_time\":\"2022-04-15T15:39:00.000Z\",\"last_edited_time\":\"2022-04-15T15:39:00.000Z\",\"created_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"last_edited_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"cover\":null,\"icon\":{\"type\":\"emoji\",\"emoji\":\"\u2603\uFE0F\"},\"parent\":{\"type\":\"page_id\",\"page_id\":\"250662d9-783d-483c-b131-341ed479f96c\"},\"archived\":false,\"properties\":{\"title\":{\"id\":\"title\",\"type\":\"title\",\"title\":[{\"type\":\"text\",\"text\":{\"content\":\"tests/test_records.py::test_page_icon
+      (setup)\",\"link\":null},\"annotations\":{\"bold\":false,\"italic\":false,\"strikethrough\":false,\"underline\":false,\"code\":false,\"color\":\"default\"},\"plain_text\":\"tests/test_records.py::test_page_icon
+      (setup)\",\"href\":null}]}},\"url\":\"https://www.notion.so/tests-test_records-py-test_page_icon-setup-746c707a24ef44f8867d17bf7d4371b7\"}"
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: '{"archived": true}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '18'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: PATCH
+    uri: https://api.notion.com/v1/pages/746c707a24ef44f8867d17bf7d4371b7
+  response:
+    content: "{\"object\":\"page\",\"id\":\"746c707a-24ef-44f8-867d-17bf7d4371b7\",\"created_time\":\"2022-04-15T15:39:00.000Z\",\"last_edited_time\":\"2022-04-15T15:39:00.000Z\",\"created_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"last_edited_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"cover\":null,\"icon\":{\"type\":\"emoji\",\"emoji\":\"\u2603\uFE0F\"},\"parent\":{\"type\":\"page_id\",\"page_id\":\"250662d9-783d-483c-b131-341ed479f96c\"},\"archived\":true,\"properties\":{\"title\":{\"id\":\"title\",\"type\":\"title\",\"title\":[{\"type\":\"text\",\"text\":{\"content\":\"tests/test_records.py::test_page_icon
+      (setup)\",\"link\":null},\"annotations\":{\"bold\":false,\"italic\":false,\"strikethrough\":false,\"underline\":false,\"code\":false,\"color\":\"default\"},\"plain_text\":\"tests/test_records.py::test_page_icon
+      (setup)\",\"href\":null}]}},\"url\":\"https://www.notion.so/tests-test_records-py-test_page_icon-setup-746c707a24ef44f8867d17bf7d4371b7\"}"
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+version: 1
