+++ tests/cassettes/test_ping_session.yaml
@@ -1,0 +1,25 @@
+interactions:
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: GET
+    uri: https://api.notion.com/v1/users/me
+  response:
+    content: '{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1","name":"SDK
+      Development","avatar_url":null,"type":"bot","bot":{"owner":{"type":"workspace","workspace":true}}}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+version: 1
