+++ tests/cassettes/test_restore_block.yaml
@@ -1,0 +1,132 @@
+interactions:
+- request:
+    body: '{"children": [{"type": "callout", "object": "block", "has_children": false,
+      "archived": false, "callout": {"text": [{"type": "text", "plain_text": "Reppearing
+      blocks!", "text": {"content": "Reppearing blocks!"}}], "color": "default"}}]}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '236'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: PATCH
+    uri: https://api.notion.com/v1/blocks/250662d9783d483cb131341ed479f96c/children
+  response:
+    content: "{\"object\":\"list\",\"results\":[{\"object\":\"block\",\"id\":\"45c7a97e-1e73-4bd8-94f9-671a92747e04\",\"created_time\":\"2022-04-15T15:38:00.000Z\",\"last_edited_time\":\"2022-04-15T15:38:00.000Z\",\"created_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"last_edited_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"has_children\":false,\"archived\":false,\"type\":\"callout\",\"callout\":{\"icon\":{\"type\":\"emoji\",\"emoji\":\"\U0001F4A1\"},\"color\":\"default\",\"text\":[{\"type\":\"text\",\"text\":{\"content\":\"Reppearing
+      blocks!\",\"link\":null},\"annotations\":{\"bold\":false,\"italic\":false,\"strikethrough\":false,\"underline\":false,\"code\":false,\"color\":\"default\"},\"plain_text\":\"Reppearing
+      blocks!\",\"href\":null}]}}],\"next_cursor\":null,\"has_more\":false}"
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: DELETE
+    uri: https://api.notion.com/v1/blocks/45c7a97e1e734bd894f9671a92747e04
+  response:
+    content: "{\"object\":\"block\",\"id\":\"45c7a97e-1e73-4bd8-94f9-671a92747e04\",\"created_time\":\"2022-04-15T15:38:00.000Z\",\"last_edited_time\":\"2022-04-15T15:38:00.000Z\",\"created_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"last_edited_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"has_children\":false,\"archived\":true,\"type\":\"callout\",\"callout\":{\"icon\":{\"type\":\"emoji\",\"emoji\":\"\U0001F4A1\"},\"color\":\"default\",\"text\":[{\"type\":\"text\",\"text\":{\"content\":\"Reppearing
+      blocks!\",\"link\":null},\"annotations\":{\"bold\":false,\"italic\":false,\"strikethrough\":false,\"underline\":false,\"code\":false,\"color\":\"default\"},\"plain_text\":\"Reppearing
+      blocks!\",\"href\":null}]}}"
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: '{"archived": false}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '19'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: PATCH
+    uri: https://api.notion.com/v1/blocks/45c7a97e1e734bd894f9671a92747e04
+  response:
+    content: "{\"object\":\"block\",\"id\":\"45c7a97e-1e73-4bd8-94f9-671a92747e04\",\"created_time\":\"2022-04-15T15:38:00.000Z\",\"last_edited_time\":\"2022-04-15T15:38:00.000Z\",\"created_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"last_edited_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"has_children\":false,\"archived\":false,\"type\":\"callout\",\"callout\":{\"icon\":{\"type\":\"emoji\",\"emoji\":\"\U0001F4A1\"},\"color\":\"default\",\"text\":[{\"type\":\"text\",\"text\":{\"content\":\"Reppearing
+      blocks!\",\"link\":null},\"annotations\":{\"bold\":false,\"italic\":false,\"strikethrough\":false,\"underline\":false,\"code\":false,\"color\":\"default\"},\"plain_text\":\"Reppearing
+      blocks!\",\"href\":null}]}}"
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: GET
+    uri: https://api.notion.com/v1/blocks/45c7a97e-1e73-4bd8-94f9-671a92747e04
+  response:
+    content: "{\"object\":\"block\",\"id\":\"45c7a97e-1e73-4bd8-94f9-671a92747e04\",\"created_time\":\"2022-04-15T15:38:00.000Z\",\"last_edited_time\":\"2022-04-15T15:38:00.000Z\",\"created_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"last_edited_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"has_children\":false,\"archived\":false,\"type\":\"callout\",\"callout\":{\"icon\":{\"type\":\"emoji\",\"emoji\":\"\U0001F4A1\"},\"color\":\"default\",\"text\":[{\"type\":\"text\",\"text\":{\"content\":\"Reppearing
+      blocks!\",\"link\":null},\"annotations\":{\"bold\":false,\"italic\":false,\"strikethrough\":false,\"underline\":false,\"code\":false,\"color\":\"default\"},\"plain_text\":\"Reppearing
+      blocks!\",\"href\":null}]}}"
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: DELETE
+    uri: https://api.notion.com/v1/blocks/45c7a97e1e734bd894f9671a92747e04
+  response:
+    content: "{\"object\":\"block\",\"id\":\"45c7a97e-1e73-4bd8-94f9-671a92747e04\",\"created_time\":\"2022-04-15T15:38:00.000Z\",\"last_edited_time\":\"2022-04-15T15:38:00.000Z\",\"created_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"last_edited_by\":{\"object\":\"user\",\"id\":\"6860cd48-ced8-41b3-b824-c7d039085db1\"},\"has_children\":false,\"archived\":true,\"type\":\"callout\",\"callout\":{\"icon\":{\"type\":\"emoji\",\"emoji\":\"\U0001F4A1\"},\"color\":\"default\",\"text\":[{\"type\":\"text\",\"text\":{\"content\":\"Reppearing
+      blocks!\",\"link\":null},\"annotations\":{\"bold\":false,\"italic\":false,\"strikethrough\":false,\"underline\":false,\"code\":false,\"color\":\"default\"},\"plain_text\":\"Reppearing
+      blocks!\",\"href\":null}]}}"
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+version: 1
