+++ tests/cassettes/test_update_block.yaml
@@ -1,0 +1,111 @@
+interactions:
+- request:
+    body: '{"children": [{"type": "to_do", "object": "block", "has_children": false,
+      "archived": false, "to_do": {"text": [{"type": "text", "plain_text": "Important
+      Task", "text": {"content": "Important Task"}}], "checked": false, "color": "default"}}]}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '242'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: PATCH
+    uri: https://api.notion.com/v1/blocks/250662d9783d483cb131341ed479f96c/children
+  response:
+    content: '{"object":"list","results":[{"object":"block","id":"5974de69-d0aa-400b-828e-d28271f93171","created_time":"2022-04-15T15:38:00.000Z","last_edited_time":"2022-04-15T15:38:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":false,"type":"to_do","to_do":{"checked":false,"color":"default","text":[{"type":"text","text":{"content":"Important
+      Task","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"Important
+      Task","href":null}]}}],"next_cursor":null,"has_more":false}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: '{"type": "to_do", "archived": false, "to_do": {"text": [{"type": "text",
+      "plain_text": "Important Task", "annotations": {"bold": false, "italic": false,
+      "strikethrough": false, "underline": false, "code": false, "color": "default"},
+      "text": {"content": "Important Task"}}], "checked": true, "color": "default"}}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '311'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: PATCH
+    uri: https://api.notion.com/v1/blocks/5974de69d0aa400b828ed28271f93171
+  response:
+    content: '{"object":"block","id":"5974de69-d0aa-400b-828e-d28271f93171","created_time":"2022-04-15T15:38:00.000Z","last_edited_time":"2022-04-15T15:38:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":false,"type":"to_do","to_do":{"checked":true,"color":"default","text":[{"type":"text","text":{"content":"Important
+      Task","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"Important
+      Task","href":null}]}}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: GET
+    uri: https://api.notion.com/v1/blocks/5974de69-d0aa-400b-828e-d28271f93171
+  response:
+    content: '{"object":"block","id":"5974de69-d0aa-400b-828e-d28271f93171","created_time":"2022-04-15T15:38:00.000Z","last_edited_time":"2022-04-15T15:38:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":false,"type":"to_do","to_do":{"checked":true,"color":"default","text":[{"type":"text","text":{"content":"Important
+      Task","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"Important
+      Task","href":null}]}}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: DELETE
+    uri: https://api.notion.com/v1/blocks/5974de69d0aa400b828ed28271f93171
+  response:
+    content: '{"object":"block","id":"5974de69-d0aa-400b-828e-d28271f93171","created_time":"2022-04-15T15:38:00.000Z","last_edited_time":"2022-04-15T15:38:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":true,"type":"to_do","to_do":{"checked":true,"color":"default","text":[{"type":"text","text":{"content":"Important
+      Task","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"Important
+      Task","href":null}]}}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+version: 1
