+++ tests/cassettes/test_update_schema.yaml
@@ -1,0 +1,115 @@
+interactions:
+- request:
+    body: '{"parent": {"type": "page_id", "page_id": "250662d9-783d-483c-b131-341ed479f96c"},
+      "title": [{"type": "text", "plain_text": "tests/test_records.py::test_update_schema
+      (setup)", "text": {"content": "tests/test_records.py::test_update_schema (setup)"}}],
+      "properties": {"Name": {"type": "title", "title": {}}}}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '308'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: POST
+    uri: https://api.notion.com/v1/databases
+  response:
+    content: '{"object":"database","id":"d9651554-755e-45f3-b57e-a68a5364d40f","cover":null,"icon":null,"created_time":"2022-04-15T15:39:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_time":"2022-04-15T15:39:00.000Z","title":[{"type":"text","text":{"content":"tests/test_records.py::test_update_schema
+      (setup)","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"tests/test_records.py::test_update_schema
+      (setup)","href":null}],"properties":{"Name":{"id":"title","name":"Name","type":"title","title":{}}},"parent":{"type":"page_id","page_id":"250662d9-783d-483c-b131-341ed479f96c"},"url":"https://www.notion.so/d9651554755e45f3b57ea68a5364d40f","archived":false}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: '{"properties": {"Name": {"type": "title", "title": {}}, "Index": {"type":
+      "number", "number": {"format": "number"}}, "Notes": {"type": "rich_text", "rich_text":
+      {}}, "Complete": {"type": "checkbox", "checkbox": {}}, "Due Date": {"type":
+      "date", "date": {}}, "Tags": {"type": "multi_select", "multi_select": {"options":
+      []}}}, "title": [{"type": "text", "plain_text": "Improved Database", "text":
+      {"content": "Improved Database"}}]}'
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      content-length:
+      - '431'
+      content-type:
+      - application/json
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: PATCH
+    uri: https://api.notion.com/v1/databases/d9651554755e45f3b57ea68a5364d40f
+  response:
+    content: '{"object":"database","id":"d9651554-755e-45f3-b57e-a68a5364d40f","cover":null,"icon":null,"created_time":"2022-04-15T15:39:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_time":"2022-04-15T15:39:00.000Z","title":[{"type":"text","text":{"content":"Improved
+      Database","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"Improved
+      Database","href":null}],"properties":{"Tags":{"id":"MvfO","name":"Tags","type":"multi_select","multi_select":{"options":[]}},"Index":{"id":"NWT%3C","name":"Index","type":"number","number":{"format":"number"}},"Complete":{"id":"N%5De%3E","name":"Complete","type":"checkbox","checkbox":{}},"Notes":{"id":"PyhA","name":"Notes","type":"rich_text","rich_text":{}},"Due
+      Date":{"id":"_GF%3C","name":"Due Date","type":"date","date":{}},"Name":{"id":"title","name":"Name","type":"title","title":{}}},"parent":{"type":"page_id","page_id":"250662d9-783d-483c-b131-341ed479f96c"},"url":"https://www.notion.so/d9651554755e45f3b57ea68a5364d40f","archived":false}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: GET
+    uri: https://api.notion.com/v1/databases/d9651554-755e-45f3-b57e-a68a5364d40f
+  response:
+    content: '{"object":"database","id":"d9651554-755e-45f3-b57e-a68a5364d40f","cover":null,"icon":null,"created_time":"2022-04-15T15:39:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_time":"2022-04-15T15:39:00.000Z","title":[{"type":"text","text":{"content":"Improved
+      Database","link":null},"annotations":{"bold":false,"italic":false,"strikethrough":false,"underline":false,"code":false,"color":"default"},"plain_text":"Improved
+      Database","href":null}],"properties":{"Tags":{"id":"MvfO","name":"Tags","type":"multi_select","multi_select":{"options":[]}},"Index":{"id":"NWT%3C","name":"Index","type":"number","number":{"format":"number"}},"Complete":{"id":"N%5De%3E","name":"Complete","type":"checkbox","checkbox":{}},"Notes":{"id":"PyhA","name":"Notes","type":"rich_text","rich_text":{}},"Due
+      Date":{"id":"_GF%3C","name":"Due Date","type":"date","date":{}},"Name":{"id":"title","name":"Name","type":"title","title":{}}},"parent":{"type":"page_id","page_id":"250662d9-783d-483c-b131-341ed479f96c"},"url":"https://www.notion.so/d9651554755e45f3b57ea68a5364d40f","archived":false}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+- request:
+    body: ''
+    headers:
+      accept:
+      - '*/*'
+      accept-encoding:
+      - gzip, deflate
+      authorization:
+      - secret...
+      connection:
+      - keep-alive
+      host:
+      - api.notion.com
+      notion-version:
+      - '2021-08-16'
+    method: DELETE
+    uri: https://api.notion.com/v1/blocks/d9651554755e45f3b57ea68a5364d40f
+  response:
+    content: '{"object":"block","id":"d9651554-755e-45f3-b57e-a68a5364d40f","created_time":"2022-04-15T15:39:00.000Z","last_edited_time":"2022-04-15T15:39:00.000Z","created_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"last_edited_by":{"object":"user","id":"6860cd48-ced8-41b3-b824-c7d039085db1"},"has_children":false,"archived":true,"type":"child_database","child_database":{"title":"Improved
+      Database"}}'
+    headers: {}
+    http_version: HTTP/1.1
+    status_code: 200
+version: 1
